Adds lower bound checks for WriteTo and ReadFrom
I'm not a fan on this solution, but it works and is in line of what the exercise asks.

I would probably create a struct which includes lower and upper and add WriteTo and ReadFrom as methods. This avoids
the hassle of passing the size_t parameter, since the method has access to the char arrays. And therefor avoids
the risk of passing a wrong size_t parameter and completely break the bound check.