Adds exercise 3_3
This made the previous linked-list of listing 3_9 and 3_10 into a doubly linked list. This was a real
bitch to implement, since there's quite some reordering to do when inserting a new element between an existing
link.

This solution isn't optimal, since it requires 2 method calls to implement the double link. I think a singular methodc
could satisfy the whole situation. But I got little experience with this data structure and requires some in-depth testing
to validate my assumptions.