Adds AccountDatabase reference to Bank (exercise 5-3)
This was actually a bit challenging, figuring out how to use arrays of user-defined types. It works, but since
I'm not an expert in this I have this constant feeling that I'm missing something or leaking memory :c

In all honesty, this isn't the cleanest of code in a long shot. I seriously despise the MakeTransfer method atm, but
it works and I kind of want to get going with the next exercise.